requirements.txt
beautifulsoup4==4.13.4
certifi==2025.4.26
charset-normalizer==3.4.2
click==8.1.8
colorama==0.4.6
gTTS==2.5.4
idna==3.10
paho-mqtt==2.1.0
PyAudio==0.2.14
pygame==2.6.1
PyYAML==6.0.2
requests==2.32.3
soupsieve==2.7
SpeechRecognition==3.14.2
typing_extensions==4.13.2
urllib3==2.4.0
